% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_function.R
\name{power_factory}
\alias{power_factory}
\title{Takes a numeric value x and returns a function which raises it's argument into the power of x}
\usage{
power_factory(x)
}
\arguments{
\item{x}{: integer}
}
\description{
Takes a numeric value x and returns a function which raises it's argument into the power of x
}
